+++ retex/renderer-base/src/main/java/com/himamis/retex/renderer/share/ScaleBox.java
@@ -67,7 +67,7 @@ public class ScaleBox extends Box {
 	}
 
 	public ScaleBox(Box b, double factor) {
-		this(b, (double) factor, (double) factor);
+		this(b, factor, factor);
 		this.factor = factor;
 	}
